+++ lib/src/theme.dart
@@ -40,6 +40,10 @@ extension WrapTheme on Widget {
       );
 }
 
+/// Handle a [Theme] and the [ThemeMode],
+/// and it will also provide the [Brightness] of current theme.
+/// And you can also modify current [Theme] and [ThemeMode] from
+/// its descendants in the widget tree via the context.
 class ThemeHandler<T extends Theme> extends StatefulWidget {
   const ThemeHandler({
     super.key,
@@ -63,19 +67,9 @@ class _ThemeHandlerState<T extends Theme> extends State<ThemeHandler<T>> {
   late T _dark = widget.dark;
   late ThemeMode _mode = widget.mode;
 
-  late T _theme = theme;
-  late Brightness _brightness = brightness;
-
-  /// Compute what theme ([T]) should be now according to [_brightness].
-  /// It's not recommended to call it directly, consider using [_theme]
-  /// to reduce unnecessary computations.
-  T get theme => _brightness == Brightness.dark ? _dark : _light;
-
   /// Compute what [Brightness] should be now according to [_mode]
   /// and [MediaQueryData.platformBrightness] of current platform.
-  /// It's not recommended to call it directly, consider using [_brightness]
-  /// to reduce unnecessary computations.
-  Brightness get brightness => _mode == ThemeMode.system
+  Brightness get adaptedBrightness => _mode == ThemeMode.system
       ? MediaQuery.of(context).platformBrightness
       : _mode == ThemeMode.dark
           ? Brightness.dark
@@ -84,59 +78,35 @@ class _ThemeHandlerState<T extends Theme> extends State<ThemeHandler<T>> {
   @override
   void didUpdateWidget(covariant ThemeHandler<T> oldWidget) {
     super.didUpdateWidget(oldWidget);
-
-    var needSetState = false;
-    if (widget.mode != _mode) needSetState = _preUpdateMode(widget.mode);
-
-    if (widget.light != _light) {
-      _light = widget.light;
-      if (_brightness == Brightness.light) needSetState = true;
-    }
-    if (widget.dark != _dark) {
-      _dark = widget.dark;
-      if (_brightness == Brightness.dark) needSetState = true;
-    }
-
-    if (needSetState) setState(() {});
+    if (_mode != widget.mode) setState(() => _mode = widget.mode);
+    if (_dark != widget.dark) setState(() => _dark = widget.dark);
+    if (_light != widget.light) setState(() => _light = widget.light);
   }
 
-  void updateMode(ThemeMode Function(ThemeMode raw) updater) {
-    if (_preUpdateMode(updater(_mode))) setState(() {});
+  void updateMode(ThemeMode mode) {
+    if (_mode != mode) setState(() => _mode = mode);
   }
 
-  bool _preUpdateMode(ThemeMode mode) {
-    if (mode == _mode) return false;
-    _mode = mode;
-    final brightness = this.brightness;
-    if (_brightness != brightness) {
-      _brightness = brightness;
-      _theme = theme;
-    }
-    return true;
-  }
-
-  void updateCurrentTheme(T Function(T raw) updater) {
-    if (_preUpdateCurrentTheme(updater(_theme))) setState(() {});
-  }
-
-  bool _preUpdateCurrentTheme(T theme) {
-    if (_theme == theme) return false;
-    switch (_brightness) {
+  void updateCurrentTheme(T theme) {
+    switch (adaptedBrightness) {
       case Brightness.dark:
-        _dark = theme;
+        if (_dark != theme) setState(() => _dark = theme);
       case Brightness.light:
-        _light = theme;
+        if (_light != theme) setState(() => _light = theme);
     }
-    return true;
   }
 
   @override
-  Widget build(BuildContext context) => widget.child
-      .foreground(context, _theme.foreground)
-      .background(_theme.background)
-      .inherit(_brightness)
-      .inherit(_theme)
-      .inherit(_mode)
-      .inherit(InheritHandlerAPI(updateMode))
-      .inherit(InheritHandlerAPI(updateCurrentTheme));
+  Widget build(BuildContext context) {
+    final brightness = adaptedBrightness;
+    final theme = adaptedBrightness == Brightness.dark ? _dark : _light;
+    return widget.child
+        .foreground(context, theme.foreground)
+        .background(theme.background)
+        .inherit(brightness)
+        .inherit(theme)
+        .inherit(_mode)
+        .inherit(InheritHandlerAPI(updateMode))
+        .inherit(InheritHandlerAPI(updateCurrentTheme));
+  }
 }